feat(frontend): Fixed PR issues
