+++ .gitignore
@@ -7,4 +7,3 @@ __pycache__
 packages/code-du-travail-data/dump.json
 package-lock.json
 yarn-error.log
-.DS_Store
