+++ packages/code-du-travail-api/src/server/routes/items/index.js
@@ -35,7 +35,6 @@ router.get("/items/:source/:slug", async ctx => {
 
   // Get current item title to find related items
   const { title } = item._source;
-  console.log("Title is :", title);
   const relatedItems = await getSearch({
     q: title,
     excludeSources: ""
